Move extensions on RuleDescription to RuleDescriptionExtensions. Do not pollute RuleDescription with responsibilities that don't belong to it.
